Move header to own component
